ad #121: refactored QrGenerator to a Service
